Fixed a problem for CPU vs CPU
I hadn't tested it before, but when the CPU played another CPU, the code failed.
Problem was that the nextPlayer could be 0, 1 or 2.
This has been changed to the first player being 0 and the second being 1.

Since in eval-simple.c the score for player X is positive added maximizing to cpusTurn(char symbol) when the player's symbol is 'X'.

This solved the problem of CPU vs CPU.

What remains is that alpha/beta plays differently from minimax.